* commit 'ef4c8a444d6558d18d3e453f7a4d080953c7c3a2 [formerly 46c440344e864d1368a2a15b2036b72f1f521b20]':
  ODY-1443: valueOf requires String! added file properties again


Former-commit-id: e95cd590788b49facc4dd1b446f1f80e79527e53

+++ common/mep/src/de/uniol/inf/is/odysseus/mep/functions/string/SubStringFunction.java
@@ -36,13 +36,13 @@ public class SubStringFunction extends AbstractFunction<String> {
 	@Override
 	public String getValue() {
 		String a = getInputValue(0);
-		Integer b = Integer.valueOf(getInputValue(1));
-		Integer c = Integer.valueOf(getInputValue(2));
+		Long b = getInputValue(1);
+		Long c = getInputValue(2);
 
 		if ((a == null) || (b == null) || (c == null)) {
 			return null;
 		}
-		return a.substring(b, c);
+		return a.substring(b.intValue(), c.intValue());
 	}
 
 }
