+++ server/memstore/de.uniol.inf.is.odysseus.memstore.mdastore/src/de/uniol/inf/is/odysseus/memstore/mdastore/functions/MDARemoveDimFunction.java
@@ -34,9 +34,9 @@ public class MDARemoveDimFunction extends AbstractFunction<Void> {
 	public Void getValue() {
 		String name = getInputValue(0);
 		Objects.requireNonNull(name);
-		Integer index = Integer.valueOf(getInputValue(1));
+		Long index = getInputValue(1);
 		Objects.requireNonNull(index);
-		removeDim(name, index);
+		removeDim(name, index.intValue());
 		return null;
 	}
